requirements.txt
psycopg2==2.9.9
psycopg2-binary==2.9.9
numpy==1.26.3
pandas==2.2.1
seaborn==0.13.2
flask==3.0
python-dotenv==1.0.1




